pyproject.toml
[tool.poetry]
name = "QHealth"
version = "0.1.0"
description = ""
authors = ["AlexJohannson <mehedsanya@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
mysqlclient = "^2.2.7"
djangorestframework-simplejwt = "^5.5.0"
celery = "^5.5.3"
redis = "^6.2.0"
django-celery-results = "^2.6.0"
django-celery-beat = "^2.8.1"
django-filter = "^25.1"
djangochannelsrestframework = "^1.3.0"
daphne = "^4.2.1"
channels-redis = "^4.3.0"
drf-yasg = "^1.21.11"
cloudinary = "^1.44.1"
fpdf = "^1.7.2"





[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

